backend/requirements.txt
flask 
psycopg2-binary
flask-cors
mysql-connector-python
python-dotenv